I have no idea why it's still not working but I realized my teams project isn't working either so I'm gonna push from laptop and pull down on my desktop and try it there just in case

+++ app/src/main/java/net/abigailthompson/grocerylist/GroceryMapActivity.java
@@ -145,8 +145,8 @@ public class GroceryMapActivity extends AppCompatActivity implements OnMapReadyC
 
             if(currentGrocery != null)
             {
-                currentGrocery.setLatitude(addresses.get(0).getLatitude());
-                currentGrocery.setLongitude(addresses.get(0).getLongitude());
+                currentGrocery.setLatitude((float) addresses.get(0).getLatitude());
+                currentGrocery.setLongitude((float) addresses.get(0).getLongitude());
             }
